fix: move dedup integration tests to smoke tests
The integration test file started a second worker via start_worker(),
which caused random flakiness when running all integration tests
locally — two workers consuming from the same queues led to stolen
tasks and missing results. CI didn't catch this because it runs test
files in isolation.

Moved the two core dedup fast-path tests to t/smoke/ which uses
pytest-celery with Docker containers for per-class worker isolation.
Updated CI workflow to remove from integration matrix and add to
smoke matrix.

Fixes local dev breakage on main reported by @Nusnus.
