+++ .github/workflows/python-package.yml
@@ -94,7 +94,6 @@ jobs:
           'test_backend.py',
           'test_canvas.py',
           'test_database_backend.py',
-          'test_dedup_chain_dispatch.py',
           'test_inspect.py',
           'test_loader.py',
           'test_mem_leak_in_exception_handling.py',
@@ -125,6 +124,7 @@ jobs:
           'test_visitor.py',
           'test_canvas.py',
           'test_consumer.py',
+          'test_dedup_chain_dispatch.py',
           'test_control.py',
           'test_signals.py',
           'test_tasks.py',
